# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an inventory management system API with authentication",
        "title": "Inventory Management System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/adjustments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of adjustments with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Adjustments"
                ],
                "summary": "List adjustments",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by reference number",
                        "name": "reference_number",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AdjustmentListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new adjustment with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Adjustments"
                ],
                "summary": "Create a new adjustment",
                "parameters": [
                    {
                        "description": "Adjustment details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateAdjustmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Adjustment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/adjustments/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific adjustment by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Adjustments"
                ],
                "summary": "Get adjustment by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Adjustment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Adjustment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing adjustment by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Adjustments"
                ],
                "summary": "Update adjustment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Adjustment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated adjustment details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateAdjustmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Adjustment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an adjustment by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Adjustments"
                ],
                "summary": "Delete adjustment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Adjustment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Adjustment": {
            "type": "object",
            "properties": {
                "adjustment_date": {
                    "type": "string"
                },
                "adjustment_reason": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "created_by_first_name": {
                    "description": "Joined fields",
                    "type": "string"
                },
                "created_by_last_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "description": "Items",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AdjustmentItem"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "processed_by": {
                    "type": "string"
                },
                "processed_by_first_name": {
                    "type": "string"
                },
                "processed_by_last_name": {
                    "type": "string"
                },
                "processed_date": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "reference_id": {
                    "type": "string"
                },
                "reference_number": {
                    "type": "string"
                },
                "reference_type": {
                    "description": "Reference fields for PO/SO connections",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_quantity": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.AdjustmentItem": {
            "type": "object",
            "properties": {
                "actual_quantity": {
                    "type": "integer"
                },
                "adjustment_id": {
                    "type": "string"
                },
                "adjustment_reason": {
                    "type": "string"
                },
                "cost_price": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "expected_quantity": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "product_name": {
                    "description": "Joined fields",
                    "type": "string"
                },
                "product_sku": {
                    "type": "string"
                },
                "quantity": {
                    "description": "Can be positive (add) or negative (subtract)",
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "reference_id": {
                    "type": "string"
                },
                "reference_number": {
                    "type": "string"
                },
                "reference_type": {
                    "description": "Reference fields for item-level tracking",
                    "type": "string"
                },
                "variance_quantity": {
                    "type": "integer"
                },
                "warehouse_id": {
                    "type": "string"
                },
                "warehouse_name": {
                    "type": "string"
                }
            }
        },
        "models.AdjustmentListResponse": {
            "type": "object",
            "properties": {
                "adjustments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Adjustment"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "models.CreateAdjustmentItemRequest": {
            "type": "object",
            "required": [
                "cost_price",
                "product_id",
                "quantity",
                "warehouse_id"
            ],
            "properties": {
                "actual_quantity": {
                    "type": "integer"
                },
                "adjustment_reason": {
                    "type": "string",
                    "enum": [
                        "receiving_discrepancy",
                        "damaged_goods",
                        "quality_issue",
                        "short_shipment",
                        "over_shipment",
                        "customer_return",
                        "defective_return",
                        "exchange",
                        "warranty_replacement",
                        "cycle_count_correction",
                        "theft_loss",
                        "expired_product",
                        "storage_damage",
                        "transfer_correction",
                        "other"
                    ]
                },
                "cost_price": {
                    "type": "number",
                    "minimum": 0
                },
                "expected_quantity": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "reference_id": {
                    "type": "string"
                },
                "reference_number": {
                    "type": "string"
                },
                "reference_type": {
                    "description": "Reference fields for item-level tracking",
                    "type": "string",
                    "enum": [
                        "purchase_order",
                        "sales_order",
                        "cycle_count",
                        "damage",
                        "theft",
                        "expired",
                        "transfer",
                        "other"
                    ]
                },
                "warehouse_id": {
                    "type": "string"
                }
            }
        },
        "models.CreateAdjustmentRequest": {
            "type": "object",
            "required": [
                "adjustment_date",
                "created_by",
                "items",
                "reference_number"
            ],
            "properties": {
                "adjustment_date": {
                    "type": "string"
                },
                "adjustment_reason": {
                    "type": "string",
                    "enum": [
                        "receiving_discrepancy",
                        "damaged_goods",
                        "quality_issue",
                        "short_shipment",
                        "over_shipment",
                        "customer_return",
                        "defective_return",
                        "exchange",
                        "warranty_replacement",
                        "cycle_count_correction",
                        "theft_loss",
                        "expired_product",
                        "storage_damage",
                        "transfer_correction",
                        "other"
                    ]
                },
                "created_by": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/models.CreateAdjustmentItemRequest"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "reference_id": {
                    "type": "string"
                },
                "reference_number": {
                    "type": "string"
                },
                "reference_type": {
                    "description": "Reference fields for PO/SO connections",
                    "type": "string",
                    "enum": [
                        "purchase_order",
                        "sales_order",
                        "cycle_count",
                        "damage",
                        "theft",
                        "expired",
                        "transfer",
                        "other"
                    ]
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "approved",
                        "completed",
                        "cancelled"
                    ]
                },
                "total_quantity": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "models.UpdateAdjustmentRequest": {
            "type": "object",
            "properties": {
                "adjustment_date": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "processed_by": {
                    "type": "string"
                },
                "processed_date": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "reference_number": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "approved",
                        "completed",
                        "cancelled"
                    ]
                },
                "total_quantity": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}